[klib] Move a helper property from `serialization.js` to `util-klib`
We want to use it in `typescript-export-standalone`, which doesn't
depend on `serialization.js`.
